[REFACTORING] Split IMAPServerPartialFetchTest
