+++ server/protocols/protocols-imap4/src/test/java/org/apache/james/imapserver/netty/IMAPServerTest.java
@@ -233,87 +233,6 @@ class IMAPServerTest {
         return ImmutableSet.of(new IpConnectionCheck());
     }
 
-    @Nested
-    class PartialFetch {
-        IMAPServer imapServer;
-        private int port;
-
-        @BeforeEach
-        void beforeEach() throws Exception {
-            imapServer = createImapServer("imapServer.xml", FetchProcessor.LocalCacheConfiguration.DEFAULT);
-            port = imapServer.getListenAddresses().get(0).getPort();
-        }
-
-        @AfterEach
-        void tearDown() {
-            imapServer.destroy();
-        }
-
-        @Test
-        void fetchShouldRetrieveMessage() throws Exception {
-            testIMAPClient.connect("127.0.0.1", port)
-                .login(USER.asString(), USER_PASS)
-                .append("INBOX", SMALL_MESSAGE);
-
-            assertThat(testIMAPClient
-                .select("INBOX")
-                .readFirstMessage())
-                .contains("* 1 FETCH (FLAGS (\\Recent \\Seen) BODY[] {21}\r\nheader: value\r\n\r\nBODY)\r\n");
-        }
-
-        @Test
-        void fetchShouldRetrieveMessageWhenOffsetAndLimitExceedingMessageSize() throws Exception {
-            testIMAPClient.connect("127.0.0.1", port)
-                .login(USER.asString(), USER_PASS)
-                .append("INBOX", SMALL_MESSAGE);
-
-            assertThat(testIMAPClient
-                .select("INBOX")
-                .readFirstMessageInMailbox("BODY[]<8.20>"))
-                .contains("* 1 FETCH (FLAGS (\\Recent \\Seen) BODY[]<8> {13}\r\nvalue\r\n\r\nBODY)\r\n");
-        }
-
-        @Test
-        void fetchShouldRetrieveMessageWhenOffsetAndLimitEqualMessageSize() throws Exception {
-            testIMAPClient.connect("127.0.0.1", port)
-                .login(USER.asString(), USER_PASS)
-                .append("INBOX", SMALL_MESSAGE);
-
-            assertThat(testIMAPClient
-                .select("INBOX")
-                .readFirstMessageInMailbox("BODY[]<8.13>"))
-                .contains("* 1 FETCH (FLAGS (\\Recent \\Seen) BODY[]<8> {13}\r\nvalue\r\n\r\nBODY)\r\n");
-        }
-
-        @Test
-        void fetchShouldRetrieveMessageWhenOffsetAndLimitBelowMessageSize() throws Exception {
-            testIMAPClient.connect("127.0.0.1", port)
-                .login(USER.asString(), USER_PASS)
-                .append("INBOX", SMALL_MESSAGE);
-
-            assertThat(testIMAPClient
-                .select("INBOX")
-                .readFirstMessageInMailbox("BODY[]<8.12>"))
-                .contains("* 1 FETCH (FLAGS (\\Recent \\Seen) BODY[]<8> {12}\r\nvalue\r\n\r\nBOD)\r\n");
-
-            assertThat(testIMAPClient
-                .select("INBOX")
-                .readFirstMessageInMailbox("BODY[]<8.12>"))
-                .contains("* 1 FETCH (BODY[]<8> {12}\r\nvalue\r\n\r\nBOD)\r\n");
-        }
-
-        @Test
-        void fetchShouldRetrieveMessageWhenOffsetAndNoLimitSpecified() throws Exception {
-            testIMAPClient.connect("127.0.0.1", port)
-                .login(USER.asString(), USER_PASS)
-                .append("INBOX", SMALL_MESSAGE);
-
-            assertThat(testIMAPClient
-                .select("INBOX")
-                .readFirstMessageInMailbox("BODY[]<8>"))
-                .contains("* 1 FETCH (FLAGS (\\Recent \\Seen) BODY[]<8> {13}\r\nvalue\r\n\r\nBODY)\r\n");
-        }
-    }
 
     @Nested
     class NoLimit {
